backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
python-multipart==0.0.9
langchain==0.1.5
pgvector==0.2.5
psycopg2-binary==2.9.9
sentence-transformers==2.6.1
python-dotenv==1.0.1
huggingface-hub==0.20.3
